requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
install==1.3.4
pyTelegramBotAPI==3.7.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
